Add vxlanInitFunc to allow for initialization that may be needed for vxlan mechanisms (#132)
In order for vppagent to properly handle vxlan tunnels in most environments, it needs to
have some kind of initialization done.  For example, in K8s, it needs to bind an AF_PACKET interface
to the Node interface.  This will vary by environment, and so shouldn't be hardcoded here, but
vxlanInitFunc allows it to be provided by the executable using the chain elements.

Signed-off-by: Ed Warnicke <hagbard@gmail.com>